+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## \[2.1.0] - Unreleased
+## \[2.1.0] - 2022-02-20
 
 * \[Feature] - Export `shellArgument` and `shellArguments` for escaping args
 * \[Feature] - Added `HelpArg` and `VerboseArg` mixin